actions: Move part into zerver.lib.users.
(cherry picked from commit ab04068294910e244f7a77e46531a142ece86b30)
