+++ zerver/management/commands/deactivate_user.py
@@ -3,9 +3,10 @@ from typing import Any
 
 from django.core.management.base import CommandError
 
-from zerver.lib.actions import do_deactivate_user, get_active_bots_owned_by_user
+from zerver.lib.actions import do_deactivate_user
 from zerver.lib.management import ZulipBaseCommand
 from zerver.lib.sessions import user_sessions
+from zerver.lib.users import get_active_bots_owned_by_user
 
 
 class Command(ZulipBaseCommand):
